0.0.1-SNAPSHOT
Whew! Added some basic core elements: Item/block system, etc.
Examples are in place for how I need to add recipes, items, etc.
This mod is “functional”, in that it can be launched but it doesn’t
really do anything. This commit is more or less for me as I finish the
day.